_OF-bot/requirements.txt
psycopg2-binary
drf-yasg
asyncpg
websockets
pybit
python-binance
okx
